WA: used int instead of int64_t

+++ CSES/Weird_Algorithm_WA.cpp
@@ -22,7 +22,7 @@ void run_case(){
 }
 
 int main(){
-  int64_t N;
+  int N;
   cin>>N;
 
   while(N!=1){
